Histogram.resize fails to clear old counts when shifting to the right.
Histogram.resize needs to shift its counts to the right when the normalized
zero index changes. However when doing so, it correctly calls arraycopy to
copy counts to the right, but forgets to erase counts between
`oldNormalizedZeroIndex` and `newNormalizedZeroIndex`. This makes the `counts`
array go out of sync with `totalCount`, which leads
`Histogram.getValueAtPercentile` to underestimate values.